requirements.txt
altgraph==0.17.4
astroid==3.3.8
autopep8==2.3.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
defusedxml==0.7.1
dill==0.3.9
filelock==3.17.0
fonttools==4.55.4
fsspec==2024.12.0
huggingface-hub==0.27.1
idna==3.10
isort==6.0.0
Jinja2==3.1.5
kiwisolver==1.4.8
lap==0.5.12
MarkupSafe==3.0.2
matplotlib==3.10.0
mccabe==0.7.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pefile==2023.2.7
pillow==11.1.0
platformdirs==4.3.6
psutil==6.1.1
py-cpuinfo==9.0.0
pycodestyle==2.12.1
pycparser==2.22
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.0
pylint==3.3.4
pyparsing==3.2.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.1
seaborn==0.13.2
setuptools==75.8.0
six==1.17.0
supervision==0.25.1
sympy==1.13.1
tomlkit==0.13.2
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.65
ultralytics-thop==2.0.14
urllib3==2.3.0